django-webservice/requirements.txt
Django==3.1.5
psycopg2==2.8.6
django-leaflet==0.27.1
Pillow==8.1.0
python-slugify==4.0.1
toml==0.10.2
django-cors-headers
django-imagekit
orjson
